refactor(embed): move render_booking_flow_html to utils namespace
Move render_html() and enqueue_frontend_assets() from
WPAppointments\Shortcode to WPAppointments\Utils\BookingFlow so the
global template tag no longer depends on shortcode internals.
