+++ plugins/wpappointments/includes/utils/booking-flow.php
@@ -1,12 +1,14 @@
 <?php
 /**
- * Booking flow attribute parsing utilities
+ * Booking flow utilities — attribute parsing and rendering
  *
  * @package WPAppointments
  * @since 0.5.0
  */
 
 namespace WPAppointments\Utils\BookingFlow;
+
+use WPAppointments\Core\PluginInfo;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -87,4 +89,76 @@ function snake_to_camel( $value ) {
 function snake_to_pascal( $value ) {
 	$parts = explode( '_', $value );
 	return implode( '', array_map( 'ucfirst', $parts ) );
+}
+
+/**
+ * Render the booking flow HTML and enqueue required assets
+ *
+ * Shared by the shortcode, widget, and template tag.
+ *
+ * @param array $attributes Booking flow attributes (camelCase keys).
+ *
+ * @return string HTML output.
+ */
+function render_html( $attributes ) {
+	enqueue_frontend_assets();
+
+	/** This filter is documented in assets/gutenberg/blocks/booking-flow/src/render.php */
+	$attributes = apply_filters( 'wpappointments_booking_flow_attributes', $attributes );
+
+	// phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.obfuscation_base64_encode -- Required for the view script to decode attributes.
+	return sprintf(
+		"<div class='wpappointments-booking-flow' data-attributes='%s'></div>",
+		esc_attr( base64_encode( wp_json_encode( $attributes ) ) )
+	);
+}
+
+/**
+ * Enqueue the frontend scripts and styles required for the booking flow
+ *
+ * When the Gutenberg block is used, these are loaded automatically via
+ * block.json viewScript/style. For shortcodes and widgets we must
+ * enqueue them manually.
+ *
+ * @return void
+ */
+function enqueue_frontend_assets() {
+	$plugin_dir  = PluginInfo::get_plugin_dir_path();
+	$plugin_url  = PluginInfo::get_plugin_dir_url();
+	$version     = PluginInfo::get_version();
+	$view_handle = 'wpappointments-booking-flow-view';
+
+	if ( wp_script_is( $view_handle, 'enqueued' ) ) {
+		return;
+	}
+
+	$asset_path = $plugin_dir . 'build/booking-flow-block-view.tsx.asset.php';
+
+	if ( ! file_exists( $asset_path ) ) {
+		return;
+	}
+
+	$asset = require $asset_path;
+
+	wp_enqueue_script(
+		$view_handle,
+		$plugin_url . 'build/booking-flow-block-view.tsx.js',
+		$asset['dependencies'],
+		$asset['version'],
+		true
+	);
+
+	wp_enqueue_style(
+		'wpappointments-frontend-css',
+		$plugin_url . 'build/frontend.tsx.css',
+		array(),
+		$version
+	);
+
+	wp_enqueue_style(
+		'wpappointments-booking-flow-view-css',
+		$plugin_url . 'build/booking-flow-block-view.tsx.css',
+		array(),
+		$version
+	);
 }
